Allow kernel arguments to begin with a hyphen
The current uhyve argument parser does not specify that flags meant for uhyve should not be parsed following the kernel argument. Therefore, any argument to the kernel beginning with a hyphen is instead treated as an argument to uhyve, throwing an error. 

In other words: 
`uhyve kernel --verbose` currently parses the "--verbose" flag, enables the verbose setting in uhyve, and executes the kernel with no arguments. There is currently no way to give an argument to the kernel if it accepts a "--verbose" flag itself. This fix makes it so that the example command will instead pass the "--verbose" flag onto the kernel.

After this fix, any flags the user means to pass to uhyve should be passed before the kernel file path, and any flags after the kernel file path are passed straight to the kernel. This is better because it matches the specification given by "uhyve --help" and gives way more flexibility to the user.