Attempt to detect Hugepages support
This commit tests if `/sys/kernel/mm/transparent_hugepage/enabled`  is present
and parses the contents.
If (transparent) hugepages are supported by the kernel, then they are
enabled by default.
If /sys/kernel/mm/transparent_hugepage/enabled is not present
the hugepages support is assumed to be not present.

I have no idea how the mac kernel works or exposes information, so a follow up commit
is probably needed to add macos support.