Add script to generate test coverage
The script uses source based coverage that was recently added to rust nightly.
It merges the .profraw files and uses grcov to generate a JSON in coveralls format, which can be used for coveralls/codecov integration.
Out of time constraints I didn't add parameter parsing, so this script currently always generates coveralls reports, and thus requires the ENV variable CODECOV_TOKEN to be set.