Added instrument feature using rftrace
Added `instrument` feature, which is disabled by default. It only works
if uhyve is compiled with `-Z instrument-mcount`. Furthermore I removed
the std::process::exit() call from uhyvelib so it is possible to write
the trace. The return value is not returned from `VirtualCPU::run()`.

Note that using rftrace may not be the best tool to trace uhyve as there
are various other tools available on a modern desktop OS, but can be
used to trace RustyHermit and using the same tool simplifies everything
a bit.