Test Coverage: Add Flag to only run Integration tests
Also add a flag to only run Integration tests.
I'm not 100% sure if this is really necessary.
Maybe we want this, since the reports will get merged and we don't want to artificially inflat the amount of times code is executed due to executing units tests multiple times. On the other hand hand that metric is probably not that valuable anyway, so I see limited use.