Replace hardcoded arch::RAM_START in BootInfo
There was an idea to initialize the mmap using different guest_address
parameters.

Despite how most of the code is already "future-proofed", with a clear
intention of heading towards that direction and `vm.rs` using
`self.mem.guest_address` everywhere (despite that value currently being
equal to zero), this change modifies a small exception to that rule.