mem.rs: Extend test_mmap_memory_readwrite()
As I was trying to understand how Uhyve interacts with the guest physical
memory, I stumbled upon this function and entered a few test values.

I believe that this change might be good to check for regressions at
best, especially because of the fact that this component is a
dependency of test_virt_to_phys.