test: improve output information
- Print to stdout as soon as the kernel successfully boots
- Print to stdout in run_simple_vm to make it clear that the kernel has
  been compiled
- Clarify the test-kernel being built, as the "name mismatch" between
  fs-test.rs and create_file.rs can be a bit confusing
- Add newline between build_hermit_bin + run_simple_vm output and
  output coming from the kernel
- Minor linguistic improvements

These changes may seem redundant to the trained eye (for example,
it is safe to assume that the output starting with "frequency: ..."
is from the kernel, thus everything that follows is from the kernel),
but can be very helpful to beginners.

Based on local changes that I created for personal use, so as to better
understand the debugging process.