docs(readme): add cargo test instructions
I came across a mysterious error that affected fs-test.

As I was experimenting with code affecting file reads and writes, I
falsely assumed that the issue was isolated to this test in particular.
I forgot that the rest of the tests would not get executed if the first
one were to fail.