test(fs-test): refactor new_file_test
new_file_test now runs a kernel called open_close_file, which attempts
to open foo.txt, write to it, then closes the file, and attempts to
open it again so as to read from it. This is more of a workaround
because of the fact that we cannot read the temporary file easily,
but it should work just as well and expand the test's coverage to
"reading the contents of a file that is a temporary file on the
host filesystem".