nit(cli): move file_mapping further up in Args
It makes sense to group it togehter with output and stats, instead of
pushing it all the way down beneath the Memory and Cpu help headings.
We presume that the settings under those headings to be more "optional"
than --file-mapping, as the defaults do not have to be changed very
often.