feat(sandbox): add fd sandbox
This extends UhyveFileMap with a sandbox mapping file descriptors to
maps, so as to prevent the opening of file descriptors that belong to
Uhyve (or to that of other VMs, when Uhyve exposes such functionality
to the user).

This change also introduces a series of test cases for the sandbox,
including close / unlink, as well as tests that deliberately use
invalid or leaked file descriptors.

Helped-by: Jonathan Klimt <jonathan.klimt@eonerc.rwth-aachen.de>