feat(sandbox): add fd sandbox
This extends UhyveFileMap with a sandbox mapping file descriptors to
maps, so as to prevent the opening of file descriptors that belong to
Uhyve (or to that of other VMs, when Uhyve exposes such functionality
to the user).

There is no Linux equivalent of unlink that is capable of removing
files using a file descriptor. We did not want to keep a file
descriptor of an unlinked file available to the VM. Therefore, we
used a BiMap that associate fds to guest paths and guest paths to fds
(instead of a simple container containing fds).

If an unlink takes place while the file is open, this would make the
subsequent close fail, as soon as the file object goes out of scope
and gets dropped. We introduced a HashSet that temporarily caches
such unlinked file descriptors.