fix(sysinfo): do not use sysinfo on x86_64
The frequency is not meant to be a frequency that the Hermit kernel
can access instead of the real-time frequency value. Hermit uses
this values as some sort of a nominal frequency for counting ticks
(see constant_tsc) internally. Having to rely on sysinfo is a
"worst-case" scenario for architectures that do not have CPUID-like
functionalities (or, if they do, we do not handle them properly
yet.)

Fixes https://github.com/hermit-os/uhyve/issues/862