feat(landlock): introduce landlock lsm support
Landlock is a stackable Linux Security Module that lets Uhyve restrict
its own capabilities (specifically: access to the filesystem) before
loading untrusted code (here: machine images, which we consider to be
"black boxes").

The Landlock module is present in newer releases of modern Linux
distributions. Uhyve's interaction with Landlock is managed by the class
`UhyveLandlockWrapper`. After Uhyve "collects" all paths it requires to
function, it enforces Landlock _before_ the image is loaded in
`UhyveVm::load_kernel`.

We determine whether a file or a directory is present at a path to:
- whitelist a non-existent file's parent directory,
  see get_parent_directory. UhyveFileMap continues to protect
  the parent directory.
- set the right permissions for paths instead of letting the
  crate "dynamically" remove directory access rights for files,
  because that causes breakage when using strict file isolation
  mode or CompatLevel::HardRequirement (see determine_ruleset).

Additional integration tests were added to check for Landlock-specific
cases, as well as unit tests of the Landlock component (which also
increase coverage of UhyveFileMap).

Uhyve's Landlock functionality is disabled on non-Linux hosts, as Landlock
is never present on non-Linux hosts. Otherwise, it is enabled by default
("CompatLevel::BestEffort", plus the requirement that a version of Landlock
(all ABIs) has to be present on the system. This goes up against Landlock's
author's opinion of taking advantage of existing sandboxing features
"opportunistically", depending on what is actually available on the operating
system. What we chose instead is to prevent execution unless if the user
explicitly opts-out using `--file-isolation none`).

For now, we rely on the serial_test crate to not break filesystem tests.
Such tests should be run in separate, newly created processes, but that was
left for later. For now, `--file-isolation strict`'s functionality is ensured
by checking against `cargo run` for now in the CI.

We check whether Landlock was already enabled once before, which
should be useful when using Uhyve as a library. If that is the case,
Uhyve returns an error, as it (probably) cannot function in conjuction
with an existing set of enforced Landlock rulesets. As soon as a
Landlock ruleset is enforced, it cannot be "relaxed".