test: make filesystem tests run serially
Chose to rely on the serial_test crate instead of manually setting the
threads that can be used for all tests to one in total, since that felt
more like a dirty hack.

This is important because of Landlock restrictions causing another
running test to fail on Linux platforms. There is some optimization
room here with serial_test's attribte macros, but they were left
out for now.