ci(landlock): alter landlock ci behavior
FileSandboxMode::Strict is used for running images (as well as
FileSandboxMode::None). On the other hand, FileSandboxMode::Normal
is used for filesystem tests.

Using "best effort" policies for filesystem tests is a bit weird,
but this is done because of weird quirks in cargo run that we have
not figured out yet.