add support of contiguous bit
The arm64 MMU supports a contiguous bit which is a hint that the TTE
is one of a set of contiguous entries which can be cached in a single
TLB entry.

The contiguous bit should be only set if the page is mapped contiguous
and each 16 * 4 KByte area have the same access property.
Consequently, the first 16 * 4 KByte area doesn't use this bit because
just the first page frame is mapped with the non-cacheable bit.