refactor: categorize Args entries
This will help us reuse existing structs by deriving the Deserialize
trait later down the line (useful for #1013 and #1015) for
configurations specific to individual machine images.

All entries under MemoryArgs, CpuArgs and GuestArgs are planned to be
part of such configurations.

Arguments in UhyveArgs should not be exposed to machine image-specific
configuration files. Assuming that these configuration files could very
well be part of the machine image at a later point (which we consider
to be untrusted), this makes the case of file_mapping particularly
noteworthy (in the sense that it is an unsolvable problem): Although
a machine image should not be able to use host filesystem locations,
it could have the ability to declare guest filesystem locations.
How this could be done is not certain yet.

On the other hand, some parameters that could, in theory, affect
isolation (e.g. MemoryArgs' no_aslr) are proposed to be part of
such a configuration file.

A concrete implementation for said configurations will follow:
For now, this is a mere refactor/clean-up/proposal that sets
certain foundations.