tests(bin): simplify Args unit tests
test_toml_are_fields_actually_skipped:
Instead of defining new variables (which was done for clarity), we
can also just use the references directly instead.

test_toml_merge:
We derive PartialEq in all *Args structs conditionally (when running
tests) and use that to compare a post-merge Args struct as expected.
This should make the effects of Merge more clear to the naked eye.
Deriving PartialEq conditionally for GuestMemorySize and CpuCore is
not possible for the tests under src/bin/uhyve.rs, because the
structs are under a different crate.