feat(gdb): choose random port when port 0 is set
Before, the user would have been able to select 0 (as allowed by
TcpListener::bind), but Uhyve would falsely report that the user
can connect to port 0 for their GDB connection.

We turned the bug into a feature. This is now documented and the
random port is correctly reported. Only useful for a very limited
subset of potential debugging-related scenarios.