feat(gdb): choose random port if no port is set
This new 'feature' is actually a workaround to a bug: Before, the user
would have been able to select 0 (as allowed by TcpListener::bind), and
this would have been fine. However, Uhyve would report that it is
awaiting a connection on port 0, when it is actually listening on
another port.

I found exposing TcpListener's functionality to the user cooler than
trying to impose restrictions, but this will presumably not be of use
to anyone, really.