fix(linux): use libc for real-time signal handling
The nix crate did not and does not support real-time signals, so we
previously transmuted said signals in the i32 representation of
signals of the nix crate. However, this stopped working after a
change in the Rust compiler, making a switch back to libc necessary
for debug mode to continue to work.

Fixes https://github.com/hermit-os/uhyve/issues/1083

Helped-by: Martin Kröning <martin.kroening@eonerc.rwth-aachen.de>
Helped-by: Ellen Emilia Anna Zscheile <fogti+devel@ytrizja.de>