refactor(kvm): derive mem regions with From trait
This commit splits the derivation of memory regions into a separate From
trait, so as to avoid doing a bunch of if-else cases inside of new_cpu
and improve readability.

This is part of an effort to reduce (future) code duplication that
was noticed during an experimental attempt of mine to port Uhyve to
aarch64 (Linux).

It could be further extended to support more than just
kvm_userspace_memory_region, but it's a start.

Helped-by: Ellen Emilia Anna Zscheile <fogti+devel@ytrizja.de>